Updates EC model to avoid spurious arithmetic underflows
